refactor: move prompts to ux and export
